response schema refactoring
