+++ src/risk_atlas_nexus/blocks/prompt_response_schema.py
