+++ src/risk_atlas_nexus/blocks/risk_detector/generic.py
@@ -4,7 +4,7 @@ from risk_atlas_nexus.ai_risk_ontology.datamodel.ai_risk_ontology import Risk
 from risk_atlas_nexus.blocks.inference import TextGenerationInferenceOutput
 from risk_atlas_nexus.blocks.prompt_templates import RISK_IDENTIFICATION_TEMPLATE
 from risk_atlas_nexus.blocks.risk_detector import RiskDetector
-from risk_atlas_nexus.blocks.inference.response_schema import LIST_OF_STR_SCHEMA
+from risk_atlas_nexus.blocks.prompt_response_schema import LIST_OF_STR_SCHEMA
 
 
 class GenericRiskDetector(RiskDetector):
