+++ src/risk_atlas_nexus/blocks/risk_mapping/risk_mapper.py
@@ -6,7 +6,7 @@ from txtai import Embeddings
 from sssom_schema import Mapping, EntityReference
 from risk_atlas_nexus.ai_risk_ontology.datamodel.ai_risk_ontology import Risk
 from risk_atlas_nexus.blocks.inference import InferenceEngine
-from risk_atlas_nexus.blocks.inference.response_schema import LIST_OF_STR_SCHEMA
+from risk_atlas_nexus.blocks.prompt_response_schema import LIST_OF_STR_SCHEMA
 from risk_atlas_nexus.blocks.prompt_templates import RISK_IDENTIFICATION_TEMPLATE
 from risk_atlas_nexus.blocks.risk_mapping import RiskMappingBase
 from risk_atlas_nexus.metadata_base import MappingMethod
